Fixes to neural.h

+++ include/neural-network.h
@@ -26,6 +26,6 @@ class NeuralNetwork {
     void setEdgeWeight(int, int, int, double);
     void gradient();
     void applyGradient(double constant);
-    void batch(int n, int** inputs, int** outputs);
+    void batch(int n, double** inputs, double** outputs);
 
 };
